feat(chat): show runtime usage in context popover
